Updated with more practise

+++ inventoryProject/src/main/java/dao/CustomerDAO.java
@@ -10,8 +10,17 @@ import java.util.ArrayList;
 import java.util.List;
 
 import classes.Customer;
+import userInput.logger;
 
-public class CustomerDAO implements DAO<Customer>{
+public class CustomerDAO implements DAO<Customer> {
+	public void logMessage(String input) {
+		logger.logMessage(input);
+	}
+
+	public void logDebug(String input) {
+		logger.logDebug(input);
+	}
+
 	@Override
 	public int countRow() {
 		int counter = 0;
@@ -26,17 +35,19 @@ public class CustomerDAO implements DAO<Customer>{
 		}
 		return counter;
 	}
+
 	@Override
 	public Customer modelFromResultSet(ResultSet resultSet) throws SQLException {
-		int ID = resultSet.getInt("ID");
 		String firstName = resultSet.getString("firstName");
 		String lastName = resultSet.getString("lastName");
 		String email = resultSet.getString("email");
 		String postcode = resultSet.getString("postcode");
 		String house = resultSet.getString("house");
-		Customer result = new Customer(ID, firstName, lastName, email, postcode, house);
+		Customer result = new Customer(firstName, lastName, email, postcode, house);
+		result.setID(resultSet.getInt("ID"));
 		return result;
 	}
+
 	@Override
 	public Customer create(Customer input) {
 		int beforeLength = countRow();
@@ -57,8 +68,9 @@ public class CustomerDAO implements DAO<Customer>{
 			return null;
 		}
 		return input;
-	
+
 	}
+
 	@Override
 	public Customer read(int id) {
 		String url = "jdbc:mysql://localhost:3306/inventory";
@@ -72,15 +84,45 @@ public class CustomerDAO implements DAO<Customer>{
 		try {
 			con = DriverManager.getConnection(url, username, password);
 			stmt = con.createStatement();
-			rs = stmt.executeQuery("Select * from customer WHERE ID=" + id);
+			rs = stmt.executeQuery("Select * from customer WHERE ID = " + id);
 			rs.next();
-			modelFromResultSet(rs);
+			result = modelFromResultSet(rs);
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
 
 		return result;
 	}
+
+	public List<Integer> getOngoing(int ID) {
+		ResultSet rs;
+		List<Integer> ongoingIDs = new ArrayList<Integer>();
+		try (Connection conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/inventory", "root",
+				"stuntflying9"); Statement statement = conn.createStatement()) {
+			rs = statement.executeQuery("SELECT * from orders WHERE customerID = " + ID);
+			while (rs.next()) {
+				if (rs.getString("status").equals("ONGOING")) {
+					logDebug("ONGOING order found with ID: " + rs.getInt("ID"));
+					ongoingIDs.add(rs.getInt("ID"));
+				}
+			}
+		} catch (SQLException e) {
+			e.printStackTrace();
+		}
+		return ongoingIDs;
+
+	}
+
+	public List<Integer> cancelAll(int ID) {
+		OrdersDAO ordDAO = new OrdersDAO();
+		List<Integer> ongoingIDs = getOngoing(ID);
+		for (int orderID : ongoingIDs) {
+			ordDAO.cancel(orderID);
+
+		}
+		return ongoingIDs;
+	}
+
 	@Override
 	public List<Customer> readAll() {
 		List<Customer> allCustomer = new ArrayList<Customer>();
@@ -102,38 +144,45 @@ public class CustomerDAO implements DAO<Customer>{
 		}
 		return allCustomer;
 	}
+
 	@Override
 	public boolean delete(int id) {
 		// need to make sure this deletes all orders and baskets
 		ResultSet rs;
 		List<Integer> orderIDs = new ArrayList<Integer>();
+		List<Integer> ongoingIDs = getOngoing(id);
 		int beforeLength = countRow();
-		try (Connection conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/inventory", "root",
-				"stuntflying9"); Statement statement = conn.createStatement()) {
-			rs = statement.executeQuery("SELECT * from orders WHERE customerID = " + id);
-			while(rs.next()) {
-			orderIDs.add(rs.getInt("ID"));
+		if (ongoingIDs.size() == 0) {
+			try (Connection conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/inventory", "root",
+					"stuntflying9"); Statement statement = conn.createStatement()) {
+				rs = statement.executeQuery("SELECT * from orders WHERE customerID = " + id);
+				while (rs.next()) {
+					orderIDs.add(rs.getInt("ID"));
+				}
+
+				for (int orderID : orderIDs) {
+					statement.executeUpdate("DELETE FROM basket where orderID = " + orderID);
+					statement.executeUpdate("DELETE FROM orders where ID = " + orderID);
+					logDebug("Deleted from database where orderID = " + orderID);
+				}
+			} catch (SQLException e) {
+				e.printStackTrace();
 			}
-			for (int orderID: orderIDs) {
-				statement.executeUpdate("DELETE * from basket where orderID = " + orderID);
-				statement.executeUpdate("DELETE * from orders where ID = " + orderID);
+			try (Connection conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/inventory", "root",
+					"stuntflying9"); Statement statement = conn.createStatement()) {
+				statement.executeUpdate("DELETE FROM customer WHERE ID = " + id);
+			} catch (SQLException e) {
+				e.printStackTrace();
 			}
-		} catch (SQLException e) {
-			e.printStackTrace();
-		}
-		try (Connection conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/inventory", "root",
-				"stuntflying9"); Statement statement = conn.createStatement()) {
-			statement.executeUpdate("DELETE FROM customer WHERE ID = " + id);
-		} catch (SQLException e) {
-			e.printStackTrace();
-		}
-		int afterLength = countRow();
-		if (beforeLength == afterLength) {
+			int afterLength = countRow();
+			if (beforeLength == afterLength) {
+				return false;
+			}
+			return true;			
+		} else
 			return false;
-		}
-		return true;
 	}
-	
+
 	public Customer update(int ID, String firstName, String lastName, String email, String postcode, String house) {
 		ResultSet rs;
 		Customer found = null;
@@ -147,6 +196,7 @@ public class CustomerDAO implements DAO<Customer>{
 			statement.setString(3, email);
 			statement.setString(4, postcode);
 			statement.setString(5, house);
+			statement.setInt(6, ID);
 			statement.executeUpdate();
 			rs = conn.createStatement().executeQuery("SELECT * from customer where ID = " + ID);
 			if (rs.next()) {
@@ -157,4 +207,4 @@ public class CustomerDAO implements DAO<Customer>{
 		}
 		return found;
 	}
-	}
+}
